Extract Pattern into pattern module
